+++ apps/cli/templates/auth/web/nuxt/app/components/SignInForm.vue
@@ -1,6 +1,5 @@
 <script setup lang="ts">
 import { z } from 'zod'
-// import { authClient } from "~/lib/auth-client";
 const {$authClient} = useNuxtApp()
 import type { FormSubmitEvent } from '#ui/types'
 
@@ -11,7 +10,7 @@ const loading = ref(false)
 
 const schema = z.object({
   email: z.string().email('Invalid email address'),
-  password: z.string().min(6, 'Password must be at least 6 characters'),
+  password: z.string().min(8, 'Password must be at least 8 characters'),
 })
 
 type Schema = z.output<typeof schema>